Change revealing of final output (#10)
* Change revealing of final output

If output is revealed to one party, revealing returns only 1 value.
This value is a correct output for the authorized party; others see
garbage.
If it is revealed to 2 parties, 2 values are returned. A non-authorized
party sees a couple of garbage values, while the authorized parties see
one garbage and one correct value.