Store rev-node-mapping in the ContextMappings (#380)
The idea of this change is to make it possible for each node of the mpc-compiled graph understand what node in the initial graph corresponds to it. It should be much easier to understand why some graphs are slow/what is the bottleneck.

There are two changes:

- In the ContextMappings struct I added rev_node_mapping field, which exactly answers the question "what node was used to generate this node in a new graph".
- Various functions, which do graph optimizations/mpc compilation now return ContextMappings instead of simple Context. If we run several optimizations in a row, we join generated ContextMappings.

There is some trickiness, because mapping is not always defined because of optimizations (e.g. we could deduplicate some nodes, or eliminate some nodes).