Run clippy on tests as well (#381)
Changed clippy command to cargo clippy --workspace --tests --all-targets --all-features and then fixed (almost automatically) all warnings.

One type of non-automatic fix. Clippy thinks that if you have some variable c, pass it to a function with c.clone() and never use it later, it is safe to remove .clone(). It is not true in our case because of Context and all weak pointers. So I changed some functions to take &Context instead of Context, so the context is still alive in the caller and also you don't need to .clone() when passing to a function.

I also don't like the needless_range_loop lint (https://rust-lang.github.io/rust-clippy/master/index.html#/needless_range_loop), so I allowed it.